// eiam-console/src/main/console-fe/src/pages/authn/IdentityProvider/data.d.ts
/*
 * eiam-console - Employee Identity and Access Management
 * Copyright © 2022-Present Jinan Yuanchuang Network Technology Co., Ltd. (support@topiam.cn)
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
/**
 * 认证源
 */
export type ListIdentityProvider = {
  id: string;
  name: string;
  desc: string;
  remark: string;
  type: any;
  enabled: boolean;
  displayed: boolean;
};

/**
 * 认证源详情
 */
export type GetIdentityProvider = {
  type: string;
  redirectUri: string;
  config: any;
};